Rename in separate commit to make both git and clippy happy
