+++ crates/ruff/src/jupyter/notebook.rs
